Move outputs into `dist/`, add `scratch-` prefix
WARNING: clients will need to update their references after this change
