+++ .gitignore
@@ -13,11 +13,11 @@ npm-*
 /.idea
 
 # Build
-/dist.js
-/vm.js
-/vm.min.js
+/dist
 /playground/assets
 /playground/media
+/playground/scratch-vm.js
+/playground/scratch-vm.js.map
 /playground/vendor.js
-/playground/vm.js
+/playground/vendor.js.map
 /playground/zenburn.css
