+++ playground/index.html
@@ -66,7 +66,7 @@
     <!-- FPS counter, Syntax highlighter, Blocks, Renderer -->
     <script src="./vendor.js"></script>
     <!-- VM Worker -->
-    <script src="./vm.js"></script>
+    <script src="./scratch-vm.js"></script>
     <!-- Playground -->
     <script src="./playground.js"></script>
     <script>
